fix: do not ignore legacy template declarations
Always throw an error if a declared template function cannot be found.
The attempt to ignore certain function names
creates more problems than it solves as it does not
incentivise to clean up templates when switching to
the newer templating library.
The test for templates with a legcacy declaration was apapted.